+++ Networking/Networking/Remote/SiteSettingsRemote.swift
@@ -12,7 +12,7 @@ public class SiteSettingsRemote: Remote {
     ///   - siteID: Site for which we'll fetch the general settings.
     ///   - completion: Closure to be executed upon completion.
     ///
-    public func loadGeneralSettings(for siteID: Int, completion: @escaping ([SiteSetting]?, Error?) -> Void) {
+    public func loadGeneralSettings(for siteID: Int64, completion: @escaping ([SiteSetting]?, Error?) -> Void) {
         let path = Constants.siteSettingsPath + Constants.generalSettingsGroup
         let request = JetpackRequest(wooApiVersion: .mark3, method: .get, siteID: siteID, path: path, parameters: nil)
         let mapper = SiteSettingsMapper(siteID: siteID, settingsGroup: SiteSettingGroup.general)
@@ -26,7 +26,7 @@ public class SiteSettingsRemote: Remote {
     ///   - siteID: Site for which we'll fetch the product settings.
     ///   - completion: Closure to be executed upon completion.
     ///
-    public func loadProductSettings(for siteID: Int, completion: @escaping ([SiteSetting]?, Error?) -> Void) {
+    public func loadProductSettings(for siteID: Int64, completion: @escaping ([SiteSetting]?, Error?) -> Void) {
         let path = Constants.siteSettingsPath + Constants.productSettingsGroup
         let request = JetpackRequest(wooApiVersion: .mark3, method: .get, siteID: siteID, path: path, parameters: nil)
         let mapper = SiteSettingsMapper(siteID: siteID, settingsGroup: SiteSettingGroup.product)
